+++ file_utils.ts
@@ -1,7 +1,15 @@
 // As of v0.55.0 this module requires the --unstable flag to be used
 import * as fs from "https://deno.land/std@v0.55.0/fs/mod.ts";
 import { Film } from "./models/film.ts";
+import { Specie } from "./models/specie.ts";
 import { IState, ModelState } from "./state.ts";
+
+/**
+ * Describes the expected structure for a .json storage.
+ */
+interface JsonStorable<T> {
+  data?: T[];
+}
 
 /**
  * Loads data and deserializes data from a json file.
@@ -13,8 +21,8 @@ async function loadDataFromFiles<T>(
   filename: string,
 ): Promise<T[]> {
   await fs.ensureDir(dataDir);
-  const result = await fs.readJson(dataDir + "/" + filename) as { data: T[] };
-  return result.data;
+  const result = await fs.readJson(dataDir + "/" + filename) as JsonStorable<T>;
+  return result.data ?? [];
 }
 
 /**
@@ -28,4 +36,17 @@ export async function createFilmStateAsync(
 ): Promise<IState<Film>> {
   const films = await loadDataFromFiles<Film>(dataDir, filmFile);
   return new ModelState<Film>(films);
+}
+
+/**
+ * Creates and seeds a ModelState for Species.
+ * @param dataDir directory holding the json file, defaults to ./data
+ * @param speciesFile json file containing species, defautls to species.json
+ */
+export async function createSpecieStateAsync(
+  dataDir: string = "./data",
+  speciesFile: string = "species.json",
+): Promise<IState<Specie>> {
+  const species = await loadDataFromFiles<Specie>(dataDir, speciesFile);
+  return new ModelState<Specie>(species);
 }