+++ models/specie.ts
@@ -1,0 +1,17 @@
+export interface Specie {
+  name: string;
+  classification: string;
+  designation: string;
+  average_height: string;
+  skin_colors: string;
+  hair_colors: string;
+  eye_colors: string;
+  average_lifespan: string;
+  homeworld: string;
+  language: string;
+  people: number[];
+  films: number[];
+  created: Date;
+  edited: Date;
+  url: number;
+}